// frontend/src/lib/generated/Incentives.types.ts
export type Uint128 = string;
export type Addr = string;
export type AssetInfo = {
  token: {
    contract_addr: Addr;
  };
} | {
  native_token: {
    denom: string;
  };
};
export interface InstantiateMsg {
  factory: string;
  guardian?: string | null;
  incentivization_fee_info?: IncentivizationFeeInfo | null;
  owner: string;
  reward_token: AssetInfo;
}
export interface IncentivizationFeeInfo {
  fee: Coin;
  fee_receiver: Addr;
}
export interface Coin {
  amount: Uint128;
  denom: string;
  [k: string]: unknown;
}
export type ExecuteMsg = {
  setup_pools: {
    pools: [string, Uint128][];
  };
} | {
  claim_rewards: {
    lp_tokens: string[];
  };
} | {
  deposit: {
    recipient?: string | null;
  };
} | {
  withdraw: {
    amount: Uint128;
    lp_token: string;
  };
} | {
  set_tokens_per_second: {
    amount: Uint128;
  };
} | {
  incentivize: {
    lp_token: string;
    schedule: InputSchedule;
  };
} | {
  incentivize_many: [string, InputSchedule][];
} | {
  remove_reward_from_pool: {
    bypass_upcoming_schedules?: boolean;
    lp_token: string;
    receiver: string;
    reward: string;
  };
} | {
  claim_orphaned_rewards: {
    limit?: number | null;
    receiver: string;
  };
} | {
  update_config: {
    generator_controller?: GeneratorControllerUpdate & string;
    guardian?: string | null;
    incentivization_fee_info?: IncentivizationFeeInfo | null;
    token_transfer_gas_limit?: number | null;
  };
} | {
  update_blocked_tokenslist: {
    add?: AssetInfo[];
    remove?: AssetInfo[];
  };
} | {
  deactivate_pool: {
    lp_token: string;
  };
} | {
  deactivate_blocked_pools: {};
} | {
  propose_new_owner: {
    expires_in: number;
    owner: string;
  };
} | {
  drop_ownership_proposal: {};
} | {
  claim_ownership: {};
};
export type GeneratorControllerUpdate = {
  set: string;
} | "unset" | "no_change";
export interface InputSchedule {
  duration_periods: number;
  reward: Asset;
}
export interface Asset {
  amount: Uint128;
  info: AssetInfo;
}
// ...
export type Decimal256 = string;
export type ArrayOfScheduleResponse = ScheduleResponse[];
export interface ScheduleResponse {
  end_ts: number;
  rps: Decimal256;
  start_ts: number;
}
export type Boolean = boolean;
export type ArrayOfString = string[];
export type ArrayOfAsset = Asset[];
export type RewardType = {
  int: AssetInfo;
} | {
  ext: {
    info: AssetInfo;
    next_update_ts: number;
  };
};
export interface PoolInfoResponse {
  last_update_ts: number;
  rewards: RewardInfo[];
  total_lp: Uint128;
}
export interface RewardInfo {
  index: Decimal256;
  orphaned: Decimal256;
  reward: RewardType;
  rps: Decimal256;
}